requirements.txt
#
# Requirements for wger for production
#

Django
django-recaptcha
reportlab
django_mobile
django-formtools
bleach
python-mimeparse
pillow
easy-thumbnails
django_compressor
icalendar
django-bootstrap-breadcrumbs
sphinx
django-sortedm2m
django-bower
invoke
requests
whitenoise
gunicorn
psycopg2-binary

# REST API
djangorestframework
django-filter
django-tastypie
django-cors-headers

# Python3 compatibility
six

# social auth
django-rest-framework-social-oauth2
